ICUI-21: pushed state onto AuthProvider and removed responsibilities on authService
**Problem:**
login/register/logout is currently in placeholder dummy state which
seems to work? Integrate service code into AuthProvider
TODO: how will I implement validate on every route/api call? is this
feasible?

**Fix:**
note to self: when the state variable belongs to a context (global), the
context provider should call the services.
when the state variable belongs to a component, the component should
call the services?

**Testing:**