ICUI-21: hooked up AuthProvider to service functions with callbacks
**Problem:**
FIXME: when the user successfully logs in/registers, go to home page?
FIXME: when the user unsuccessfully logs in/registers, display error
message?
FIXME: decode user's JWT token and read their user login name
FIXME: when the user is logged in, do not allow user to navigate to
/register and /login
FIXME: when the user is logged in, hide the register and login buttons
FIXME: when the user is not logged in, hide the locked function buttons

**Fix:**
kinda realized that Promises are async lolool

**Testing:**